+++ vscge/include/vscge/event/key_event.h
@@ -23,115 +23,17 @@
 #include "vscge/event/base_event.h"
 
 namespace vs {
-
-// TODO(Victor): Move to platform.
-enum class Key {
-  kEscape = VK_ESCAPE,
-
-  kLeft  = VK_LEFT,
-  kUp    = VK_UP,
-  kRight = VK_RIGHT,
-  kDown  = VK_DOWN,
-
-  kBack = VK_BACK,
-  kTab  = VK_TAB,
-
-  kReturn  = VK_RETURN,
-  kShift   = VK_SHIFT,
-  kControl = VK_CONTROL,
-
-  kClear = VK_CLEAR,
-  kMenu  = VK_MENU,
-  kPause = VK_PAUSE,
-
-  kSpace   = VK_SPACE,
-  kEnd     = VK_END,
-  kHome    = VK_HOME,
-  kSelect  = VK_SELECT,
-  kPrint   = VK_PRINT,
-  kExecute = VK_EXECUTE,
-  kInsert  = VK_INSERT,
-  kDelete  = VK_DELETE,
-
-  k0 = 0x30,
-  k1 = 0x31,
-  k2 = 0x32,
-  k3 = 0x33,
-  k4 = 0x34,
-  k5 = 0x35,
-  k6 = 0x36,
-  k7 = 0x37,
-  k8 = 0x38,
-  k9 = 0x39,
-
-  kA = 0x41,
-  kB = 0x42,
-  kC = 0x43,
-  kD = 0x44,
-  kE = 0x45,
-  kF = 0x46,
-  kG = 0x47,
-  kH = 0x48,
-  kI = 0x49,
-  kJ = 0x4A,
-  kK = 0x4B,
-  kL = 0x4C,
-  kM = 0x4D,
-  kN = 0x4E,
-  kO = 0x4F,
-  kP = 0x50,
-  kQ = 0x51,
-  kR = 0x52,
-  kS = 0x53,
-  kT = 0x54,
-  kU = 0x55,
-  kV = 0x56,
-  kW = 0x57,
-  kX = 0x58,
-  kY = 0x59,
-  kZ = 0x5A,
-
-  kNumpad0   = VK_NUMPAD0,
-  kNumpad1   = VK_NUMPAD1,
-  kNumpad2   = VK_NUMPAD2,
-  kNumpad3   = VK_NUMPAD3,
-  kNumpad4   = VK_NUMPAD4,
-  kNumpad5   = VK_NUMPAD5,
-  kNumpad6   = VK_NUMPAD6,
-  kNumpad7   = VK_NUMPAD7,
-  kNumpad8   = VK_NUMPAD8,
-  kNumpad9   = VK_NUMPAD9,
-  kMultiply  = VK_MULTIPLY,
-  kAdd       = VK_ADD,
-  kSeparator = VK_SEPARATOR,
-  kSubtract  = VK_SUBTRACT,
-  kDecimal   = VK_DECIMAL,
-  kDivide    = VK_DIVIDE,
-
-  kF1  = VK_F1,
-  kF2  = VK_F2,
-  kF3  = VK_F3,
-  kF4  = VK_F4,
-  kF5  = VK_F5,
-  kF6  = VK_F6,
-  kF7  = VK_F7,
-  kF8  = VK_F8,
-  kF9  = VK_F9,
-  kF10 = VK_F10,
-  kF11 = VK_F11,
-  kF12 = VK_F12,
-
-  kNumlock = VK_NUMLOCK,
-  kScroll  = VK_SCROLL,
-};
+namespace platform {
+enum class Key;
+}  // namespace platform
 
 struct VS_API KeyEvent : public Event {
   KeyEvent() = default;
-  explicit KeyEvent(Key key) : key(key) {}
+  explicit KeyEvent(platform::Key key) : key(key) {}
 
   virtual ~KeyEvent() = default;
 
-  Key key;
+  platform::Key key;
 
   static constexpr EventType TypeStatic() { return EventType::kKey; }
   std::string TypeName() const override { return "KeyEvent"; }