+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Some more matrix types.
+
+### Changed
+
+- Platform layer is now inside the `platform` namespace
+
 ## [0.13.4] - 2020-03-30
 
 ### Added